moved test harness class and fixed test errors
